[cogito-mobile] Moves components into their own folders
