+++ workspaces/cogito-mobile/home/Home.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Text, View } from 'react-native'
 import { Navigation } from 'react-native-navigation'
-import styles from './Styles'
+import styles from '../Styles'
 
 export class Home extends React.Component {
   static options () {
