+++ workspaces/cogito-mobile/home/index.js
@@ -1,0 +1,1 @@
+export { Home } from './Home'
