+++ workspaces/cogito-mobile/identity-manager/IdentityManager.test.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { render } from 'react-native-testing-library'
 import { IdentityManager } from './IdentityManager'
 import { Navigation } from 'react-native-navigation'
-import { CreateIdentity } from './CreateIdentity'
+import { CreateIdentity } from './create/CreateIdentity'
 
 it('can add an identity', () => {
   render(<IdentityManager />)
