+++ workspaces/cogito-mobile/identity-manager/create/Layout.ios.js
@@ -1,4 +1,4 @@
-import { toolbarButton } from './ToolbarButton'
+import { toolbarButton } from '../../navigation/ToolbarButton'
 
 export const layout = {
   topBar: {
