+++ workspaces/cogito-mobile/identity-manager/index.js
@@ -1,0 +1,2 @@
+export { IdentityManager } from './IdentityManager'
+export { CreateIdentity } from './create/CreateIdentity'
