+++ workspaces/cogito-mobile/index.js
@@ -1,7 +1,6 @@
 import { Navigation } from 'react-native-navigation'
-import { Home } from './Home'
-import { IdentityManager } from './IdentityManager'
-import { CreateIdentity } from './CreateIdentity'
+import { Home } from './home'
+import { IdentityManager, CreateIdentity } from './identity-manager'
 
 Navigation.registerComponent('Home', () => Home)
 Navigation.registerComponent('IdentityManager', () => IdentityManager)
